Minor update - handle 3rd party connections
This update will listen for 3rd party IP connections and disconnect for a period configured in the config.ini file. The script will then attempt to reconnect. Useful if you still want to use the web/smartphone apps. Note that it takes about 20 seconds to disconnect after which you can try to connect to the alarm again.